refactor(client): update file names, folder structure
