+++ client/components/index.js
@@ -1,6 +1,6 @@
 // Export hub for React components
-export { default as Navbar } from './navbar'
-export { default as Login } from './login-form'
-export { default as SignUp } from './signUp-form'
-export { default as UserHome } from './user-home'
+export { default as Navbar } from './Navbar'
+export { default as Login } from './user/LoginForm'
+export { default as SignUp } from './user/SignUpForm'
+export { default as Landing } from './landing/LandingContainer'
